// service/src/helpers/playlist.helper.ts
import { createImageLinks } from './link.helper';
import { createArtistMapPayload } from './artist.helper';
import { createSongPayload } from './song.helper';
import type { playlistAPIResponseSchema, playlistSchema } from '@/schemas/playlist.schema';
import type { z } from 'zod';

export const createPlaylistPayload = (
	playlist: z.infer<typeof playlistAPIResponseSchema>,
): z.infer<typeof playlistSchema> => ({
	id: playlist.id,
	name: playlist.title,
	description: playlist.header_desc,
	type: playlist.type,
	year: playlist.year ? Number(playlist.year) : null,
	playCount: playlist.play_count ? Number(playlist.play_count) : null,
	language: playlist.language,
	explicitContent: playlist.explicit_content === '1',
	url: playlist.perma_url,
	songCount: playlist.list_count ? Number(playlist.list_count) : null,
	artists: playlist.more_info.artists?.map(createArtistMapPayload) || null,
	image: createImageLinks(playlist.image),
	songs: (playlist.list && playlist.list?.map(createSongPayload)) || null,
});
